MW-40 #comment tweaked delete functionality and have written unit and integration tests

+++ src/main/java/com/qa/choonz/service/TrackService.java
@@ -76,8 +76,13 @@ public class TrackService {
 	}
 
 	public boolean delete(long id) {
+		if(!this.repo.existsById(id)) {
+			throw new TrackNotFoundException();
+		}
 		this.repo.deleteById(id);
-		return !this.repo.existsById(id);
+		boolean exists = this.repo.existsById(id);
+		
+		return !exists;
 	}
 
 }
